backend/requirements.txt
python-dotenv
pydantic
email-validator
numpy
httpx
asyncpg
cachetools
fastapi
uvicorn

starlette
PyJWT
aiogram
prometheus-client
prometheus-fastapi-instrumentator